ingest-data/pyproject.toml
[tool.poetry]
name = "ingest-data"
version = "1.0.0"
description = "Ingest data from Google cloud storage to BigQuery"
authors = ["Marc Jamot <marc@getstoix.com>"]
license = "MIT"
packages = [
    { include = "./**/*.py" }
]

[tool.poetry.dependencies]
python = ">=3.6.2, <3.10"
google-cloud-bigquery = "^2.23.3"
google-cloud-storage = "^1.41.1"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
pytest = "^6.2.4"
tox = "^3.24.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
